Use standard github URLs instead of flakehub
The * URLs don't work well in some situations. For instance, they
cause this command to error:

```
nix develop github:biobricks-ai/dev-shell
error:
       … while updating the lock file of flake 'github:biobricks-ai/dev-shell/c61ec5a4d3449226ab545371e14b31e62c7eb6b6'

       error: cannot write modified lock file of flake 'github:biobricks-ai/dev-shell' (use '--no-write-lock-file' to ignore)
```

They also require a recent Nix version, which is not always present.

The flakehub URLs only provide minimal convenience that is not worth
the complications. I will be sticking with github URLs in the future.